init : fix param .

+++ interfaces/innerkits/reboot/init_reboot.c
@@ -46,7 +46,7 @@ int DoReboot(const char *cmdContent)
         }
         return 0;
     }
-    int length = strlen(cmdContent);
+    size_t length = strlen(cmdContent);
     if (length > MAX_REBOOT_VAUE_SIZE) {
         INIT_LOGE("DoReboot api error, cmdContent = %s, length = %d.", cmdContent, length);
         return -1;
