+++ ueventd/ueventd_utils.c
@@ -77,6 +77,6 @@ int StringToInt(const char *str, int defaultValue)
         return defaultValue;
     }
     errno = 0;
-    int value = strtoul(str, NULL, DECIMALISM);
+    int value = (int)strtoul(str, NULL, DECIMALISM);
     return errno != 0 ? defaultValue : value;
 }
